Separate Account and AccountPermissions extensions
